refactor: move arquivos para aumentar coerencia do codigo
