+++ web/src/data/index.js
@@ -1,5 +1,7 @@
 import { initializeFirebase } from './firebase';
+import { config } from 'dotenv';
 
 export const InitializeDataLayer = () => {
+  config();
   initializeFirebase()
 };
