+++ web/src/domain/index.js
@@ -2,6 +2,7 @@ import {
   ReduxAdapter,
   AuthenticationInteractor,
   CounterInteractor,
+  configureStore,
 } from 'core';
 import {
   sendPasswordResetEmail,
@@ -19,4 +20,8 @@ export const InitializeDomainLayer = () => {
     authStateListener: () => ({}),
   });
   ReduxAdapter.counter = new CounterInteractor(-5, 5);
+
+  const store = configureStore();
+
+  return { store }
 };
