+++ web/src/presentation/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { App } from './App';
+import * as ServiceWorker from './serviceWorker';
+
+export const InitializePresentationLayer = (store) => {
+  ReactDOM.render(<App store={store} />, document.getElementById('react-app'));
+
+  // If you want your app to work offline and load faster, you can change
+  // unregister() to register() below. Note this comes with some pitfalls.
+  // Learn more about service workers: http://bit.ly/CRA-PWA
+  ServiceWorker.unregister();
+};
